OneDrive/Escritorio/Salud argentina/pyproject.toml
[project]
name = "saludos-argentina"
version = "0.1.0"
description = "SaludOS Argentina — infraestructura de salud digital B2B"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "redis[hiredis]>=5.2.0",
    "celery[redis]>=5.4.0",
    "httpx>=0.28.0",
    "zeep>=4.3.0",
    "tenacity>=9.0.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-passlib>=1.7.7",
    "respx>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"
markers = [
    "integration: marks tests that require real external credentials (skipped by default)",
]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
